move repetitive functions in the utils/fonctions.ts
